New hook for both filtering and sorting
